requirements.txt
fastapi
uvicorn
streamlit
openai
google-generativeai
pillow
pywin32
requests
python-multipart
python-docx
PyPDF2

